refactor(distributed): unify platform collectives module
