+++ rl_engine/distributed/__init__.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (c) 2026 RL-Kernel Contributors
 
-from rl_engine.distributed.collectives import DeterministicCollective
-from rl_engine.distributed.transport_collectives import (
+from rl_engine.distributed.collectives import (
+    DeterministicCollective,
     RCCLDeterministicCollective,
     TorchDistributedDeterministicCollective,
     create_deterministic_collective,
